+++ msal/src/main/java/com/microsoft/identity/nativeauth/statemachine/states/SignInStates.kt
@@ -53,9 +53,6 @@ import com.microsoft.identity.nativeauth.statemachine.errors.SignInContinuationE
 import com.microsoft.identity.nativeauth.statemachine.errors.SignInErrorTypes
 import com.microsoft.identity.nativeauth.statemachine.errors.SignInSubmitPasswordError
 import com.microsoft.identity.nativeauth.statemachine.errors.SubmitCodeError
-import com.microsoft.identity.nativeauth.statemachine.results.SignInMFAGetAuthMethodsResult
-import com.microsoft.identity.nativeauth.statemachine.results.SignInMFARequiredResult
-import com.microsoft.identity.nativeauth.statemachine.results.SignInMFASubmitChallengeResult
 import com.microsoft.identity.nativeauth.statemachine.results.SignInResendCodeResult
 import com.microsoft.identity.nativeauth.statemachine.results.SignInResult
 import com.microsoft.identity.nativeauth.statemachine.results.SignInSubmitCodeResult
@@ -328,199 +325,6 @@ class SignInCodeRequiredState internal constructor(
         }
 
         override fun newArray(size: Int): Array<SignInCodeRequiredState?> {
-            return arrayOfNulls(size)
-        }
-    }
-}
-
-class SignInAwaitingMFAState(
-    override val continuationToken: String,
-    override val correlationId: String,
-    private val scopes: List<String>?,
-    private val config: NativeAuthPublicClientApplicationConfiguration
-) : BaseState(continuationToken = continuationToken, correlationId = correlationId), State, Parcelable {
-
-    // Challenge default auth method
-    suspend fun sendChallenge(): SignInMFARequiredResult {
-        return withContext(Dispatchers.IO) {
-            // if /challenge returns HTTP 200
-            if (true) {
-                SignInMFARequiredResult.VerificationRequired(
-                    nextState = SignInMFARequiredState(
-                        continuationToken = continuationToken,
-                        correlationId = correlationId,
-                        scopes = scopes,
-                        config = config
-                    ),
-                    codeLength = 6,
-                    sentTo = "user@contoso.com",
-                    channel = "email"
-                )
-            } else {
-                // /challenge returns introspect_required
-                // call /introspect and return authMethods
-                SignInMFARequiredResult.SelectionRequired(
-                    nextState = SignInMFARequiredState(
-                        continuationToken = continuationToken,
-                        correlationId = correlationId,
-                        scopes = scopes,
-                        config = config
-                    ),
-                    authMethods = listOf(1, 2, 3)
-                )
-            }
-        }
-    }
-
-    constructor(parcel: Parcel) : this(
-        continuationToken = parcel.readString()  ?: "",
-        correlationId = parcel.readString() ?: "UNSET",
-        scopes = parcel.createStringArrayList(),
-        config = parcel.serializable<NativeAuthPublicClientApplicationConfiguration>() as NativeAuthPublicClientApplicationConfiguration
-    )
-
-    override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeString(continuationToken)
-        parcel.writeString(correlationId)
-        parcel.writeStringList(scopes)
-        parcel.writeSerializable(config)
-    }
-
-    override fun describeContents(): Int {
-        return 0
-    }
-
-    companion object CREATOR : Parcelable.Creator<SignInAwaitingMFAState> {
-        override fun createFromParcel(parcel: Parcel): SignInAwaitingMFAState {
-            return SignInAwaitingMFAState(parcel)
-        }
-
-        override fun newArray(size: Int): Array<SignInAwaitingMFAState?> {
-            return arrayOfNulls(size)
-        }
-    }
-}
-
-class SignInMFARequiredState(
-    override val continuationToken: String,
-    override val correlationId: String,
-    private val scopes: List<String>?,
-    private val config: NativeAuthPublicClientApplicationConfiguration
-) : BaseState(continuationToken = continuationToken, correlationId = correlationId), State, Parcelable {
-
-    // Challenge default auth method
-    suspend fun sendChallenge(): SignInMFARequiredResult {
-        return withContext(Dispatchers.IO) {
-            // if /challenge returns HTTP 200
-            if (true) {
-                SignInMFARequiredResult.VerificationRequired(
-                    nextState = SignInMFARequiredState(
-                        continuationToken = continuationToken,
-                        correlationId = correlationId,
-                        scopes = scopes,
-                        config = config
-                    ),
-                    codeLength = 6,
-                    sentTo = "user@contoso.com",
-                    channel = "email"
-                )
-            } else {
-                // /challenge returns introspect_required
-                // call /introspect and return authMethods
-                SignInMFARequiredResult.SelectionRequired(
-                    nextState = SignInMFARequiredState(
-                        continuationToken = continuationToken,
-                        correlationId = correlationId,
-                        scopes = scopes,
-                        config = config
-                    ),
-                    authMethods = listOf(1, 2, 3)
-                )
-            }
-        }
-    }
-
-    // Challenge specified auth methods
-    suspend fun sendChallenge(authMethod: Int): SignInMFARequiredResult {
-        return withContext(Dispatchers.IO) {
-            // if /challenge returns HTTP 200
-            if (true) {
-                SignInMFARequiredResult.VerificationRequired(
-                    nextState = SignInMFARequiredState(
-                        continuationToken = continuationToken,
-                        correlationId = correlationId,
-                        scopes = scopes,
-                        config = config
-                    ),
-                    codeLength = 6,
-                    sentTo = "user@contoso.com",
-                    channel = "email"
-                )
-            } else {
-                // /challenge returns introspect_required
-                // call /introspect and return authMethods
-                SignInMFARequiredResult.SelectionRequired(
-                    nextState = SignInMFARequiredState(
-                        continuationToken = continuationToken,
-                        correlationId = correlationId,
-                        scopes = scopes,
-                        config = config
-                    ),
-                    authMethods = listOf(1, 2, 3)
-                )
-            }
-        }
-    }
-
-    // Call /introspect
-    suspend fun getAuthMethods(): SignInMFAGetAuthMethodsResult {
-        return SignInMFAGetAuthMethodsResult(
-            authMethods = listOf(1, 2, 3)
-        )
-    }
-
-    // Call /token
-    suspend fun submitChallenge(code: Int): SignInMFASubmitChallengeResult {
-        // If /token returns HTTP 200
-        return if (true) {
-            SignInResult.DummyComplete()
-        } else {
-            // If /token returns another mfa_required error
-            SignInResult.MFARequired(
-               nextState = SignInAwaitingMFAState(
-                   continuationToken = continuationToken,
-                   correlationId = correlationId,
-                   scopes = scopes,
-                   config = config
-               )
-            )
-        }
-    }
-
-    constructor(parcel: Parcel) : this(
-        continuationToken = parcel.readString()  ?: "",
-        correlationId = parcel.readString() ?: "UNSET",
-        scopes = parcel.createStringArrayList(),
-        config = parcel.serializable<NativeAuthPublicClientApplicationConfiguration>() as NativeAuthPublicClientApplicationConfiguration
-    )
-
-    override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeString(continuationToken)
-        parcel.writeString(correlationId)
-        parcel.writeStringList(scopes)
-        parcel.writeSerializable(config)
-    }
-
-    override fun describeContents(): Int {
-        return 0
-    }
-
-    companion object CREATOR : Parcelable.Creator<SignInMFARequiredState> {
-        override fun createFromParcel(parcel: Parcel): SignInMFARequiredState {
-            return SignInMFARequiredState(parcel)
-        }
-
-        override fun newArray(size: Int): Array<SignInMFARequiredState?> {
             return arrayOfNulls(size)
         }
     }
