+++ msal/src/test/java/com/microsoft/identity/client/e2e/tests/network/nativeauth/SignInTest.kt
@@ -23,13 +23,14 @@
 
 package com.microsoft.identity.client.e2e.tests.network.nativeauth
 
+import com.microsoft.identity.client.e2e.utils.assertState
 import com.microsoft.identity.internal.testutils.nativeauth.ConfigType
-import com.microsoft.identity.nativeauth.statemachine.errors.MFAError
 import com.microsoft.identity.nativeauth.statemachine.errors.SignInError
-import com.microsoft.identity.nativeauth.statemachine.results.SignInMFARequiredResult
+import com.microsoft.identity.nativeauth.statemachine.results.MFARequiredResult
 import com.microsoft.identity.nativeauth.statemachine.results.SignInResult
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert
+import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertTrue
 import org.junit.Test
 
@@ -59,51 +60,24 @@ class SignInTest : NativeAuthPublicClientApplicationAbstractTest() {
     @Test
     fun testSignInMFASimple() = runTest {
         val result = application.signIn("user", "password".toCharArray())
-        when (result) {
-            is SignInResult.Complete -> {
+        assertState<SignInResult.MFARequired>(result)
+        // Initiate challenge, send code to email
+        val sendChallengeResult = (result as SignInResult.MFARequired).nextState.sendChallenge()
+        assertState<MFARequiredResult.VerificationRequired>(sendChallengeResult)
+        (sendChallengeResult as MFARequiredResult.VerificationRequired)
+        assertNotNull(sendChallengeResult.sentTo)
+        assertNotNull(sendChallengeResult.codeLength)
+        assertNotNull(sendChallengeResult.channel)
 
-            }
-            is SignInResult.MFARequired -> {
-                val nextState = result.nextState
-                // Initiate challenge, send code to email
-                val sendChallengeResult = nextState.sendChallenge()
-                when (sendChallengeResult) {
-                    is SignInMFARequiredResult.VerificationRequired -> {
-                        // Show code UI
-                        showCodeUI(
-                            sentTo = sendChallengeResult.sentTo,
-                            codeLength = sendChallengeResult.codeLength
-                        )
+        // Retrieve all methods to build additional "pick MFA method UI"
+        val authMethodsResult = sendChallengeResult.nextState.getAuthMethods()
+        assertTrue(authMethodsResult.authMethods.isNotEmpty())
+        // call /challenge with specified ID
+        val sendChallengeResult2 = sendChallengeResult.nextState.sendChallenge(authMethodsResult.authMethods[0])
+        assertState<MFARequiredResult.VerificationRequired>(sendChallengeResult2)
 
-                        // Retrieve all methods to build additional "pick MFA method UI"
-                        val authMethodsResult = sendChallengeResult.nextState.getAuthMethods()
-                        // call /challenge with specified ID
-                        sendChallengeResult.nextState.sendChallenge(authMethodsResult.authMethods[0])
-                    }
-                    is SignInMFARequiredResult.SelectionRequired -> {
-                        val authMethods = sendChallengeResult.authMethods
-                        val sendSelectedChallengeResult = sendChallengeResult.nextState.sendChallenge(authMethods[0])
-                        if (sendSelectedChallengeResult is SignInMFARequiredResult.VerificationRequired) {
-                            val submitCodeResult = sendSelectedChallengeResult.nextState.submitChallenge(1234)
-                            if (submitCodeResult is SignInResult.DummyComplete) {
-                                assertTrue(true)
-                            } else {
-                                assertTrue(false)
-                            }
-                        }
-                    }
-                    is MFAError -> {
-                        assertTrue(false)
-                    }
-                }
-            }
-            is SignInError -> {
-                assertTrue(false)
-            }
-        }
-    }
-
-    private fun showCodeUI(sentTo: String, codeLength: Int) {
-
+        // Submit the user supplied code to the API
+        val submitCodeResult = (sendChallengeResult2 as MFARequiredResult.VerificationRequired).nextState.submitChallenge(1234)
+        assertState<SignInResult.DummyComplete>(submitCodeResult)
     }
 }
